refactor: extract auth test methods to auth test fixture
